foodgram/requirements.txt
Django==3.2.16
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
python-dotenv==1.0.0
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
djoser==2.1.0
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
webcolors==1.11.1
regex==2023.6.3
python-decouple==3.8